Keep agent drop routing out of Workspace.swift
The new pane-level agent routing helpers are split into a focused Workspace extension so the fix does not grow the already-budgeted Workspace.swift file. This preserves the routing behavior while satisfying the CI file-length guard.

Constraint: workflow-guard-tests enforces a per-file Swift line budget for Workspace.swift
Rejected: Refresh the file-length budget | this change can avoid adding new debt by moving cohesive routing helpers to a small extension
Confidence: high
Scope-risk: narrow
Tested: git diff --check; Workspace.swift line count is 13973 against CI budget 13975
Not-tested: Local tests and local build per task policy
